chore: add UI improvements
